Make rendering of pluggable search controls more generic
